codegenerator/requirements.txt
# Core dependencies
streamlit>=1.31.0
openai>=1.12.0
python-dotenv>=1.0.0
typing_extensions>=4.5.0
